requirements.txt
matplotlib==3.4.3
opencv_contrib_python==4.5.5.64
opencv_python==4.5.5.64
opencv_python_headless==4.5.5.62
Pillow==9.2.0
SimpleITK==2.2.1
evalutils==0.3.1
scikit-learn>=0.21
scipy>=1.2.1
fastai==1.0.61
tqdm
torchvision>=0.10.0
torch>=1.6.0
numpy